+++ packages/terra-folder-tree/src/subcomponents/FolderTreeItem.jsx
@@ -5,7 +5,7 @@ import { injectIntl } from 'react-intl';
 
 import Spacer from 'terra-spacer';
 import Arrange from 'terra-arrange';
-import { IconFolder } from 'terra-icon';
+import { IconFolder, IconCaretRight, IconCaretDown } from 'terra-icon';
 import ThemeContext from 'terra-theme-context';
 
 import styles from './FolderTreeItem.module.scss';
@@ -26,6 +26,10 @@ const propTypes = {
    */
   subfolderItems: PropTypes.arrayOf(PropTypes.element),
   /**
+   * Whether or not the item is expanded. Only items with subfolderItems can be expanded.
+   */
+  isExpanded: PropTypes.bool,
+  /**
    * Whether or not the item is selected. Since this component has the appearance of a radio button group, only one item should be selected at a time.
    */
   isSelected: PropTypes.bool,
@@ -33,6 +37,10 @@ const propTypes = {
    * Callback function for click event.
    */
   onClick: PropTypes.func,
+  /**
+   * Callback function for expand/collapse toggle event.
+   */
+  onToggle: PropTypes.func,
   /**
    * @private
    * Level of nesting for this item.
@@ -46,25 +54,30 @@ const propTypes = {
 };
 
 const defaultProps = {
+  isExpanded: false,
   isSelected: false,
   level: 0,
 };
 
 const FolderTreeItem = ({
   icon,
-  label,
-  subfolderItems,
+  isExpanded,
   isSelected,
-  onClick,
+  label,
   level,
+  onClick,
+  onToggle,
+  subfolderItems,
   intl,
 }) => {
   const theme = useContext(ThemeContext);
+  const isFolder = subfolderItems?.length > 0;
 
-  const subfolder = subfolderItems?.length > 0 ? (
+  const subfolder = isFolder ? (
     <ul
       className={cx('subfolder')}
       role="group"
+      hidden={!isExpanded}
     >
       {subfolderItems.map((item) => (
         <FolderTreeItem
@@ -77,6 +90,9 @@ const FolderTreeItem = ({
   ) : null;
 
   const itemIcon = subfolder ? <IconFolder a11yLabel={intl.formatMessage({ id: 'Terra.folder-tree.folder-icon' })} /> : icon;
+  const expandCollapseIcon = isExpanded
+    ? <IconCaretDown height="8px" width="8px" style={{ verticalAlign: 'baseline' }} /> // eslint-disable-line react/forbid-component-props
+    : <IconCaretRight height="8px" width="8px" style={{ verticalAlign: 'baseline' }} />; // eslint-disable-line react/forbid-component-props
 
   const itemClassNames = classNames(
     cx(
@@ -86,27 +102,51 @@ const FolderTreeItem = ({
     ),
   );
 
+  const handleToggle = (event) => {
+    if (event.target.nodeName === 'INPUT') {
+      return;
+    }
+
+    if (onToggle) {
+      onToggle();
+    }
+  };
+
   return (
     <>
+      {/* TODO: Re-enable this eslint rule once keyboard handling is implemented */}
+      {/* eslint-disable-next-line jsx-a11y/click-events-have-key-events */}
       <li
         className={itemClassNames}
         role="treeitem"
+        aria-expanded={isFolder ? isExpanded : null}
         aria-selected={isSelected}
+        onClick={isFolder ? handleToggle : onClick}
       >
         <input
           type="radio"
           checked={isSelected}
-          onClick={onClick}
+          onChange={onClick}
+          aria-hidden // Hiding the radio button from assistive technology since they cannot be grouped correctly
+          tabIndex={-1} // Prevent tabbing to the button since it should not be read or acknowledged by assistive technology
+          className={cx('radio')}
         />
         {/* eslint-disable-next-line react/forbid-dom-props */}
         <span style={{ paddingLeft: `${level}rem` }}>
           <Arrange
             fitStart={(
               <Spacer paddingLeft="medium" paddingRight="medium" isInlineBlock>
+                {
+                  isFolder ? (
+                    <Spacer paddingRight="small" isInlineBlock>
+                      {expandCollapseIcon}
+                    </Spacer>
+                  ) : null
+                }
                 {itemIcon}
               </Spacer>
             )}
-            fill={label}
+            fill={<span>{label}</span>}
             alignFitStart="center"
           />
         </span>