+++ packages/terra-framework-docs/src/terra-dev-site/test/embedded-content-consumer/consumers/DataStatusConsumer.test.jsx
@@ -46,7 +46,9 @@ const DataStatusConsumer = () => (
     <h2>Embedded Content</h2>
     <p>The following is an embedded content within an iframe.</p>
     <EmbeddedContentConsumer
-      src="../../../../#/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/data-status-provider"
+      // Use `../../../../` prefix if there is any changes to the providers file so we can see changes during the PR-preview build.
+      // Otherwise, keep the existing `/terra-framework/` prefix so the page will load when viewing the live site under /terra-ui and /terra-framework.
+      src="/terra-framework/#/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/data-status-provider"
       className={cx('iframe')}
       onMount={onMount}
       onLaunch={onLaunch}