+++ packages/terra-data-grid/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+* Added
+  * Added keyboard navigation support for sections.
+
+* Fixed
+  * Removed unnecessary rerenders
+  
+## 1.6.0 - (December 5, 2023)
+
 * Changed
   * Removed console warning message when no pinned columns exist.