+++ packages/terra-menu/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+* Fixed
+  * Fixed the issue with `contentWidth` as auto value.
+  
+* Changed
+  * Allow menu/sub-menu to dynamically set height based on menu items.
+
 ## 6.81.0 - (November 23, 2023)
 
 * Changed