+++ packages/terra-table/CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## Unreleased
 
-* Breaking Changes
+* Fixed
+  * Removed unnecessary rerenders
+
+## 5.3.0 - (December 6, 2023)
+
+* Breaking Changes (Note: this breaking change is needed at this time but does not have major impact)
   * Updated the onRowSelect callback to return an object containing section and row ids.
   
 * Added
@@ -10,6 +15,7 @@
   
 * Changed
   * Removed console warning message when no pinned columns exist.
+  * Updated the table component to properly support selection callbacks for the grid context.
 
 ## 5.2.1 - (December 1, 2023)